fix: security & correctness hardening (external review — v0.8.1) (#18)
* fix: security & correctness hardening (external review — v0.8.1)

Fixes all findings from an external review (2xP0, 5xP1, 2xP2), each verified
against the code first; the two P0s and the deadlock were reproduced before
fixing. Verification record: docs/security/REVIEW-2026-07.md.

P0 — stored XSS could steal the browser signing key. Untrusted pack labels
were interpolated into innerHTML. All RuleGarden rendering rebuilt on DOM
createElement + textContent (an el() helper); zero innerHTML in the UI; a
strict Content-Security-Policy (script-src 'self' 'wasm-unsafe-eval').
Proven inert in-browser against HTML/quote/SVG/event-handler payloads; a
dependency-free CI check (web/rulegarden/security.test.mjs) guards it.

P0 — a non-canonical hypervector crashed the engine. Hex/binary decoders
accepted the final word's unused 48 bits; one set bit indexed the
[10000]uint32 counter array out of range (reproduced from pack JSON).
core.Hypervector.ValidateCanonical REJECTS (never masks — masking would
alter signed bytes) at hex decode, pack materialization, the memory-image
loader, and certificate state. Fuzz tests for pack and memory parsers.

P1 — live sync sent the brain before identity approval. Replaced with a
mutual-approval handshake (hello -> local approval -> accept -> both ready);
nothing is sent or applied before both approve, and a keyed peer's packs
must be signed by that key. Verified in-browser.

P1 — opposite-direction concurrent Merge could deadlock. Merge now snapshots
the source ledger under a single read lock, releases it, then takes only the
destination lock. Regression: 200x A.Merge(B)||B.Merge(A) + fan-in, -race.

P1 — World.Hash() omitted future-relevant state (bound vectors, vocab seed,
writer site/seq). Now hashes a complete LocalStateFingerprint per creature.

P1 — "generalization" could be reported with no identifiable source; now
instinct in both Brain.Decide and the certificate's DerivePolicyOutcome.
Contributors carry Hamming Distance; certificates record GeneralizationRadius
so a verifier re-derives a generalization from named sources.

P1 — RuleGarden lesson labels were an unforgeable source of truth only by
convention. apply_pack now rejects any imported lesson whose bound vector
contradicts the percept/action its label claims (forged label/vector pair).

P2 — remaining JS-unsafe uint64s (PackEntry.Seq, world Seed, cert VocabSeed,
state seed) use one shared quoted-string wire type (core.QuotedU64); a test
drives a real JSON number->float64->JSON cycle.

P2 — MaxStmtStream did not bound AST traversal (ast.Inspect prunes children,
not siblings). Replaced with a manual visitor and a hard global node budget.

Also: engine.Version 0.3.0-dev -> 0.8.1 (it is signed into every cert;
existing receipts still verify via their own recorded version). CI gains a
Go 1.22 minimum-version job and a web UI build + security-check job.


* ci: keep required check names stable, add Go 1.22 as a separate job

The 'test (<os>)' job names are required contexts in the branch-protection ruleset; adding the go version to the matrix name orphaned them and blocked merges. Revert the rename and add the 1.22 minimum-version coverage as its own additive job.


---------

+++ rulegarden/sign.go
@@ -27,7 +27,7 @@ func (p *Pack) CanonicalBytes() []byte {
 		Seed    uint64  `json:"seed"`
 		Ticks   int     `json:"ticks"`
 		Events  []Event `json:"events"`
-	}{p.Version, p.Seed, p.Ticks, p.Events}
+	}{p.Version, uint64(p.Seed), p.Ticks, p.Events}
 	b, err := json.Marshal(shadow)
 	if err != nil {
 		return nil // plain-data structs cannot fail to marshal; guard defensively
